Restore embedded_hal trait, make new API default, add examples
- Bring back embedded_hal::i2c::I2c impl on I2cMasterDevice with
  assert_eq on address mismatch
- New I2C API is the default; legacy available via i2c-legacy feature
- Rename old examples to *_legacy (gated behind i2c-legacy)
- Add new i2c_ssd1306 and i2c_master_slave examples using new API
- All examples use cfg-gated mod example pattern matching rmt examples
